pyproject.toml
[project]
name = "wix_stripe_github_access"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sebastian Blanco Abrajan", email = "sebastian.blanco@kaxanuk.mx"},
]
dependencies = [
    "flask>=3.1.2",
    "stripe>=13.0.1",
    "python-dotenv>=1.1.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
